Remove DynamoDBIngestStatusStoreResources
