+++ java/cdk/src/main/java/sleeper/cdk/stack/ingest/IngestStatusStoreResources.java
@@ -34,15 +34,7 @@ public interface IngestStatusStoreResources {
 
     static IngestStatusStoreResources from(Construct scope, String id, InstanceProperties properties, ManagedPoliciesStack policiesStack) {
         if (properties.getBoolean(INGEST_STATUS_STORE_ENABLED)) {
-            return new IngestStatusStoreStack(scope, id, properties, policiesStack).getResources();
-        } else {
-            return none();
-        }
-    }
-
-    static IngestStatusStoreResources from(Construct scope, InstanceProperties properties, ManagedPoliciesStack policiesStack) {
-        if (properties.getBoolean(INGEST_STATUS_STORE_ENABLED)) {
-            return new DynamoDBIngestStatusStoreResources(scope, properties, policiesStack);
+            return new IngestStatusStoreStack(scope, id, properties, policiesStack);
         } else {
             return none();
         }
